+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 0.3.0
 
+- Switch to using new template compiler for Angular 5+.
 - Add `--skip <regex>` option, to filter out projects
   permitted by `--match <regex>`.